fixing deepsource issues
