+++ src/Spinner.Test/Helper/Interceptors/StringModelWithInterceptor.cs
@@ -1,0 +1,10 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test.Helper.Interceptors
+{
+    public class StringModelWithInterceptor
+    {
+        [ReadProperty(0, 20, typeof(ObjectToStringInterceptor))]
+        public string Value { get; set; }
+    }
+}
