+++ src/Spinner.Test/Helper/Models/TypeModels/ByteModel.cs
@@ -1,0 +1,10 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test.Helper.Models.TypeModels
+{
+    public class ByteModel
+    {
+        [ReadProperty(0, 3)]
+        public byte Value { get; set; }
+    }
+}
