+++ src/Spinner.Test/Helper/Models/TypeModels/DateTimeModel.cs
@@ -1,0 +1,10 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test.Helper.Models.TypeModels
+{
+    public class DateTimeModel
+    {
+        [ReadProperty(0, 19)]
+        public System.DateTime Value { get; set; }
+    }
+}
