+++ src/Spinner.Test/Helper/Models/TypeModels/DoubleModel.cs
@@ -1,0 +1,10 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test.Helper.Models.TypeModels
+{
+    public class DoubleModel
+    {
+        [ReadProperty(0, 15)]
+        public double Value { get; set; }
+    }
+}
