+++ src/Spinner.Test/Helper/Models/TypeModels/Int32Model.cs
@@ -1,0 +1,10 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test.Helper.Models.TypeModels
+{
+    public class Int32Model
+    {
+        [ReadProperty(0, 10)]
+        public int Value { get; set; }
+    }
+}
