+++ src/Spinner.Test/Helper/Models/TypeModels/NIntModel.cs
@@ -1,0 +1,10 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test.Helper.Models.TypeModels
+{
+    public class NIntModel
+    {
+        [ReadProperty(0, 10)]
+        public nint Value { get; set; }
+    }
+}
