+++ src/Spinner.Test/Helper/Models/TypeModels/NUIntModel.cs
@@ -1,0 +1,10 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test.Helper.Models.TypeModels
+{
+    public class NUIntModel
+    {
+        [ReadProperty(0, 10)]
+        public nuint Value { get; set; }
+    }
+}
