+++ src/Spinner.Test/Helper/Models/TypeModels/SByteModel.cs
@@ -1,0 +1,10 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test.Helper.Models.TypeModels
+{
+    public class SByteModel
+    {
+        [ReadProperty(0, 4)]
+        public sbyte Value { get; set; }
+    }
+}
