+++ src/Spinner.Test/Helper/Models/TypeModels/SingleModel.cs
@@ -1,0 +1,10 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test.Helper.Models.TypeModels
+{
+    public class SingleModel
+    {
+        [ReadProperty(0, 10)]
+        public float Value { get; set; }
+    }
+}
