+++ src/Spinner.Test/Helper/Models/TypeModels/StringModel.cs
@@ -1,0 +1,10 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test.Helper.Models.TypeModels
+{
+    public class StringModel
+    {
+        [ReadProperty(0, 20)]
+        public string Value { get; set; }
+    }
+}
