+++ src/Spinner.Test/Helper/Models/TypeModels/TimeSpanModel.cs
@@ -1,0 +1,10 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test.Helper.Models.TypeModels
+{
+    public class TimeSpanModel
+    {
+        [ReadProperty(0, 20)]
+        public System.TimeSpan Value { get; set; }
+    }
+}
