+++ src/Spinner.Test/Helper/Models/TypeModels/UInt32Model.cs
@@ -1,0 +1,10 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test.Helper.Models.TypeModels
+{
+    public class UInt32Model
+    {
+        [ReadProperty(0, 10)]
+        public uint Value { get; set; }
+    }
+}
