+++ src/Spinner.Test/Helper/Models/TypeModels/UInt64Model.cs
@@ -1,0 +1,10 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test.Helper.Models.TypeModels
+{
+    public class UInt64Model
+    {
+        [ReadProperty(0, 20)]
+        public ulong Value { get; set; }
+    }
+}
